# perf(common): faster random string generator