--- a/packages/core/scanner.ts
+++ b/packages/core/scanner.ts
@@ -362,9 +362,7 @@ export class DependenciesScanner {
     if (!providersKeys.includes(type as string)) {
       return this.container.addProvider(provider as any, token);
     }
-    const providerToken = `${
-      type as string
-    } (UUID: ${randomStringGenerator()})`;
+    const providerToken = `${type as string} (UID: ${randomStringGenerator()})`;
 
     let scope = (provider as ClassProvider | FactoryProvider).scope;
     if (isNil(scope) && (provider as ClassProvider).useClass) {